Remove OpenCensus support for metrics
